ssjgl

ssjgl

Usage

ssjgl(
  Y,
  penalty = "fused",
  lambda0,
  lambda1,
  lambda2,
  v1 = 1,
  v0s = seq(1e-04, 0.01, len = 10),
  doubly = FALSE,
  rho = 1,
  a = 1,
  b = 1,
  maxitr.em = 500,
  tol.em = 1e-04,
  maxitr.jgl = 500,
  tol.jgl = 1e-05,
  warm = NULL,
  warm.connected = NULL,
  truncate = 1e-05,
  normalize = FALSE,
  c = 0.1,
  impute = TRUE
)

Arguments

Y

List of k data matrices

penalty

Either "fused" or "group"

lambda0

scalar for penalization of the diagonals

lambda1

either constant or matrix of values to search over

lambda2

either constant or matrix of values to search over

v1

edgewise penalties

v0s

edgewise penalties

doubly

True or False

rho

default as 1

a

initializing parameters

b

initializing parameters

maxitr.em

max iterations of EM algorithm. Default 500

tol.em

default 1e-4 ADD MORE

maxitr.jgl

max iterations for JGL. default 500

tol.jgl

default 1e-4 ADD MORE

warm

default NULL. warming parameter for M step

warm.connected

parameter for M-step

truncate

cutoff to truncate default 1e-5

normalize

True or False

c

constant. Default 0.1

impute

true or false

Value

similar output to JGL